Move inner class
